requirements.txt
asgiref==3.2.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.3
django-appconf==1.0.3
django-compressor==2.4
django-sass-processor==0.8
django-webpack-loader==0.7.0
jedi==0.16.0
libsass==0.19.4
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
asgiref==3.2.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.3
django-appconf==1.0.3
django-compressor==2.4
django-filter==2.2.0
django-sass-processor==0.8
django-webpack-loader==0.7.0
djangorestframework==3.11.0
ipdb==0.13.2
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
libsass==0.19.4
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
git+git://github.com/michdol/django-subdomains.git#egg=django-subdomains
Markdown==3.2.1
django-cors-headers==3.2.1
pyyaml==5.3